Refactor module_canvas into renderer and controller submodules
- Extract `controller.rs` for state logic
- Extract `renderer.rs` for UI rendering
- Move `NodeTemplateTrait` impl to `types.rs`
- Update `mod.rs` to delegate
